Fixed handling of Markdown headers in comments.
The trick is to mark sections as "multiline comment"
(in the "meta" tag) after the inline comment marker
has been stripped once.

This means the marker won't be stripped again, which
solves the problem of markdown headers in python comments.

The solution is not exactly "beautiful", because seting
th section as "multiline comment" is slightly misleading,
but maybe you cold change the "multiline comment" string
to something like "dont strip inline", or something more
explicit.